Use size_t instead of int in SpdyToHttpFilter
r=bmcquade

+++ src/mod_spdy/apache/filters/spdy_to_http_filter.h
@@ -68,7 +68,7 @@ class SpdyToHttpFilter {
   std::string data_buffer_;
   HttpStringBuilder visitor_;
   SpdyToHttpConverter converter_;
-  int next_read_start_;
+  size_t next_read_start_;
 
   DISALLOW_COPY_AND_ASSIGN(SpdyToHttpFilter);
 };
